added polygon 2D and stratified sampling
